+++ Sources/BraintreeCore/Braintree-Version.h
@@ -1,1 +1,1 @@
-#define BRAINTREE_VERSION (@"5.17.0")
+#define BRAINTREE_VERSION (@"5.18.0")